# docs/03_openapi_contract_draft.yaml
openapi: 3.1.0
info:
  title: Encrypted Data Query & Management API
  version: 0.1.0-draft
  description: |
    基于 SRS v1.0 的接口合同草案。
    所有接口默认返回 application/json，除文件上传外。
  contact:
    name: Project Team
servers:
  - url: /
    description: Local deployment
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前登录用户
      responses:
        '200':
          description: 当前用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/data/import:
    post:
      tags: [DataImport]
      summary: 上传并创建导入任务（管理员及以上）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: 导入任务已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/data/import/{jobId}:
    get:
      tags: [DataImport]
      summary: 查询导入任务状态
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 导入任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidParam:
              value:
                code: INVALID_ARGUMENT
                message: 请求参数不合法
                trace_id: 67b8d10c5f1948f8
    Unauthorized:
      description: 未登录或令牌无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: UNAUTHORIZED
                message: 认证失败
                trace_id: 8e4fbeebdbd54d44
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                code: FORBIDDEN
                message: 无权限访问该资源
                trace_id: 3a2ee53fabff4d56
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: NOT_FOUND
                message: 资源不存在
                trace_id: 6f460f54b6f84607
  schemas:
    Role:
      type: string
      enum: [SUPER_ADMIN, ADMIN, USER]
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 128
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          nullable: true
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: access token 有效期（秒）
          example: 3600
        user:
          $ref: '#/components/schemas/UserSummary'
...
    UserSummary:
      type: object
      required: [id, username, role, is_active]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        is_active:
          type: boolean
...
    UserProfileResponse:
      type: object
      required: [data]
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: object
              properties:
                last_login_at:
                  type: string
                  format: date-time
                  nullable: true
                quota:
                  $ref: '#/components/schemas/Quota'
...
    Quota:
      type: object
      properties:
        daily_limit:
          type: integer
        daily_used:
          type: integer
        total_limit:
          type: integer
        total_used:
          type: integer
...
    ImportJob:
      type: object
      required:
        - id
        - filename
        - status
        - total_rows
        - success_rows
        - skipped_rows
        - failed_rows
      properties:
        id:
          type: integer
        filename:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, FAILED, CANCELLED]
        total_rows:
          type: integer
        success_rows:
          type: integer
        skipped_rows:
          type: integer
        failed_rows:
          type: integer
        error_summary:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
    ImportJobResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ImportJob'
...
    ErrorResponse:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          enum:
            - INVALID_ARGUMENT
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - QUOTA_EXCEEDED
            - IMPORT_FAILED
            - INTERNAL_ERROR
        message:
          type: string
        trace_id:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true